requirements.txt
pyyaml
argcomplete
pybtex